Ver1.5.0 增加字体颜色与不透明度选项
