+++ StickyNotes/StickyNotes/App.xaml
