+++ StickyNotes/StickyNotes/App.xaml.cs
