+++ StickyNotes/StickyNotes/AssemblyInfo.cs
