+++ StickyNotes/StickyNotes/ColorPickerDialog.xaml
@@ -1,10 +1,14 @@
 ﻿<Window x:Class="StickyNotes.ColorPickerDialog"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        Title="选择颜色" 
-        Width="300" 
-        Height="300"
-        WindowStartupLocation="CenterOwner" ResizeMode="NoResize" SizeToContent="Manual">
+        Title="选择颜色"
+        Width="420"
+        Height="460"
+        MinWidth="360"
+        MinHeight="300"
+        WindowStartupLocation="CenterOwner"
+        ResizeMode="CanResizeWithGrip"
+        SizeToContent="Manual">
     <Grid Margin="10">
         <StackPanel>
             <StackPanel Orientation="Horizontal">
@@ -19,6 +23,25 @@
                 <TextBlock Text="蓝色：" VerticalAlignment="Center"/>
                 <Slider x:Name="BlueSlider" Minimum="0" Maximum="255" Width="200" ValueChanged="Slider_ValueChanged"/>
             </StackPanel>
+            <StackPanel Orientation="Horizontal">
+                <TextBlock Text="不透明度：" VerticalAlignment="Center"/>
+                <Slider x:Name="AlphaSlider" Minimum="0" Maximum="255" Width="200" ValueChanged="Slider_ValueChanged"/>
+            </StackPanel>
+
+            <TextBlock Text="字体颜色：" Margin="0,8,0,2"/>
+            <StackPanel Orientation="Horizontal">
+                <TextBlock Text="红色：" VerticalAlignment="Center"/>
+                <Slider x:Name="FontRedSlider" Minimum="0" Maximum="255" Width="170" ValueChanged="Slider_ValueChanged"/>
+            </StackPanel>
+            <StackPanel Orientation="Horizontal">
+                <TextBlock Text="绿色：" VerticalAlignment="Center"/>
+                <Slider x:Name="FontGreenSlider" Minimum="0" Maximum="255" Width="170" ValueChanged="Slider_ValueChanged"/>
+            </StackPanel>
+            <StackPanel Orientation="Horizontal">
+                <TextBlock Text="蓝色：" VerticalAlignment="Center"/>
+                <Slider x:Name="FontBlueSlider" Minimum="0" Maximum="255" Width="170" ValueChanged="Slider_ValueChanged"/>
+            </StackPanel>
+
             <StackPanel Orientation="Horizontal" Margin="0,10">
                 <Border Width="30" Height="30" Margin="2" Cursor="Hand"
                         Background="#FFFF0000" Tag="255,0,0" BorderBrush="Black" BorderThickness="1"
@@ -36,8 +59,13 @@
                         Background="White" Tag="255, 255, 255" BorderBrush="Black" BorderThickness="1"
                         MouseLeftButtonDown="ColorBlock_MouseLeftButtonDown"/>
             </StackPanel>
-            <Border Height="50" BorderBrush="Gray" BorderThickness="1"
-                    Background="{Binding PreviewColor}"/>
+
+            <Border Height="60" BorderBrush="Gray" BorderThickness="1"
+                    Background="{Binding PreviewColor}" Padding="6">
+                <TextBlock Text="预览文字" VerticalAlignment="Center" HorizontalAlignment="Left"
+                           FontSize="16" Foreground="{Binding PreviewTextBrush}" />
+            </Border>
+
             <StackPanel Orientation="Horizontal" HorizontalAlignment="Right" Margin="0,10">
                 <Button Content="确定" Width="80" Click="OkButton_Click"/>
                 <Button Content="取消" Width="80" Margin="5,0,0,0" Click="CancelButton_Click"/>
