+++ StickyNotes/StickyNotes/Converters/ColorToBrushConverter.cs
