+++ StickyNotes/StickyNotes/MainWindow.xaml
