+++ StickyNotes/StickyNotes/StickyNotes.csproj
@@ -1,11 +1,11 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-	  <Version>1.4.5</Version>
+	  <Version>1.5.0</Version>
 	  <!-- 主版本号 -->
-	  <AssemblyVersion>1.4.5.0</AssemblyVersion>
+	  <AssemblyVersion>1.5.0.0</AssemblyVersion>
 	  <!-- 程序集版本 -->
-	  <FileVersion>1.4.5.0</FileVersion>
+	  <FileVersion>1.5.0.0</FileVersion>
 	  <!-- 文件版本 -->
     <OutputType>WinExe</OutputType>
 	<TargetFramework>net8.0-windows</TargetFramework>
