+++ StickyNotes/StickyNotes/Utils/ScreenHelper.cs
