+++ StickyNotes/StickyNotes/Win32ApiHelper.cs
