+++ StickyNotes/StickyNotes/app.manifest
