+++ StickyNotes/StickyNotes/icon/icon.png
