--- StickyNotes_Ai/README.md
@@ -1,20 +1,0 @@
-AI 修改打包 - StickyNotes (仅包含已更改的文件)
-
-说明：此文件夹包含对 StickyNotes 项目所做的 AI 修改的源文件副本
-
-包含的文件：
-- StickyNotes\ColorPickerDialog.xaml (颜色选择对话框：增加不透明度滑块、字体颜色滑块、可调整窗口大小、添加预览文字)
-- StickyNotes\ColorPickerDialog.xaml.cs (逻辑：新增 Alpha/FontColor 处理、PreviewTextBrush、SelectedFontColor 返回)
-- StickyNotes\Controls\StickyNoteControl.xaml (便签控件：将 TextBlock.Foreground 绑定到 FontColor)
-- StickyNotes\Controls\StickyNoteControl.xaml.cs (便签控件后端：新增 FontColor 属性并应用)
-- StickyNotes\MainWindow.xaml.cs (主窗口：传递 SelectedFontColor，创建/加载/保存笔记时持久化 FontColor)
-- StickyNotes\StickyNotes.csproj (项目文件：与原项目一致的副本，便于编译)
-
-AI 修改要点：
-- 新增背景颜色的不透明度（Alpha）滑块，颜色使用 Color.FromArgb 保存并预览。
-- 在颜色选择窗口加入字体颜色（R/G/B 滑块），并新增“预览文字”用于实时预览字体颜色。
-- 将所选字体颜色通过 SelectedFontColor 返回，并在新建便签时应用到便签文本上。
-- 保存/加载时同时持久化便签的 FontColor，以保证重启后颜色不丢失。
-- 调整颜色选择对话框为可调整大小，避免控件被遮挡。
-
-此文件也是由AI生成总结
