--- StickyNotes_Ai/StickyNotes/Controls/StickyNoteControl.xaml
@@ -1,78 +1,0 @@
-<Window x:Class="StickyNotes.Controls.StickyNoteControl"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        WindowStyle="None"
-        AllowsTransparency="True"
-        Background="Transparent"
-        Topmost="True"
-        SizeToContent="WidthAndHeight"
-        ShowInTaskbar="False">
-    <Border Background="{Binding BackgroundColor, 
-                    Converter={StaticResource ColorToBrushConverter},
-                    RelativeSource={RelativeSource AncestorType=Window}}" CornerRadius="5" 
-           BorderBrush="#666" BorderThickness="1"
-           Padding="10" MouseDown="Border_MouseDown" CacheMode="BitmapCache">
-        <Border.ContextMenu>
-            <ContextMenu>
-                <MenuItem Header="删除" Click="DeleteMenuItem_Click"/>
-                <MenuItem x:Name="PinMenuItem" Header="固定到窗口" Click="PinToWindow_Click"/>
-                <MenuItem x:Name="UnpinMenuItem" Header="取消固定" Click="Unpin_Click" Visibility="Collapsed"/>
-            </ContextMenu>
-        </Border.ContextMenu>
-        <Border.Triggers>
-            <EventTrigger RoutedEvent="MouseEnter">
-                <BeginStoryboard>
-                    <Storyboard>
-                        <DoubleAnimation 
-                    Storyboard.TargetProperty="Opacity" 
-                    To="0.9" Duration="0:0:0.2"/>
-                    </Storyboard>
-                </BeginStoryboard>
-            </EventTrigger>
-            <EventTrigger RoutedEvent="MouseLeave">
-                <BeginStoryboard>
-                    <Storyboard>
-                        <DoubleAnimation 
-                    Storyboard.TargetProperty="Opacity" 
-                    To="1" Duration="0:0:0.2"/>
-                    </Storyboard>
-                </BeginStoryboard>
-            </EventTrigger>
-        </Border.Triggers>
-        <Border.RenderTransform>
-            <TranslateTransform x:Name="MoveTransform"/>
-        </Border.RenderTransform>
-        <StackPanel>
-            <TextBox x:Name="EditBox"
-                 Visibility="Collapsed"
-                 FontSize="{Binding FontSize}"
-                 Text="{Binding NoteContent, UpdateSourceTrigger=PropertyChanged}"
-                 AcceptsReturn="True"
-                 TextWrapping="Wrap"
-                 Padding="5"
-                 LostFocus="EditBox_LostFocus"
-                 KeyDown="EditBox_KeyDown">
-                <TextBox.ContextMenu>
-                    <ContextMenu>
-                        <MenuItem Header="加粗" Click="BoldMenuItem_Click"/>
-                        <MenuItem Header="下划线" Click="UnderlineMenuItem_Click"/>
-                        <MenuItem Header="删除线" Click="StrikethroughMenuItem_Click"/>
-                        <Separator/>
-
-                        <MenuItem Header="剪切" Command="Cut"/>
-                        <MenuItem Header="复制" Command="Copy"/>
-                        <MenuItem Header="粘贴" Command="Paste"/>
-                        <Separator/>
-                        <MenuItem Header="全选" Command="SelectAll"/>
-                    </ContextMenu>
-                </TextBox.ContextMenu>
-            </TextBox>
-            <TextBlock FontSize="{Binding FontSize}" x:Name="DisplayTextBlock"
-                Style="{StaticResource NoteTextStyle}"
-                Text="{Binding NoteContent}"
-                MouseDown="TextBlock_MouseDown"
-                TextWrapping="Wrap"
-                Foreground="{Binding FontColor, Converter={StaticResource ColorToBrushConverter}, RelativeSource={RelativeSource AncestorType=Window}}"/>
-        </StackPanel>
-    </Border>
-</Window>
