--- StickyNotes_Ai/StickyNotes/StickyNotes.csproj
@@ -1,48 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-	  <Version>1.4.5</Version>
-	  <!-- 主版本号 -->
-	  <AssemblyVersion>1.4.5.0</AssemblyVersion>
-	  <!-- 程序集版本 -->
-	  <FileVersion>1.4.5.0</FileVersion>
-    <OutputType>WinExe</OutputType>
-	<TargetFramework>net8.0-windows</TargetFramework>
-	<IncludeAllContentForSelfExtract>true</IncludeAllContentForSelfExtract>
-    <Nullable>enable</Nullable>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <UseWPF>true</UseWPF>
-	<UseWindowsForms>true</UseWindowsForms>
-	<ApplicationManifest>app.manifest</ApplicationManifest>
-	<PackageIcon></PackageIcon>
-	<PackageReadmeFile></PackageReadmeFile>
-	<ApplicationIcon>icon\r7ats-wjgih-001.ico</ApplicationIcon>
-  </PropertyGroup>
-  <ItemGroup>
-    <None Remove="Fonts\1.woff2" />
-  </ItemGroup>
-
-
-
-  <ItemGroup>
-    <Content Include="icon\r7ats-wjgih-001.ico">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </Content>
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Resource Include="Fonts\1.woff2" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Update="icon\icon.png">
-      <Pack>True</Pack>
-      <PackagePath>\</PackagePath>
-    </None>
-  </ItemGroup>
-
-</Project>
